requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
numpy==2.2.2
apache-airflow-providers-common-io==1.6.0
geopy==2.4.1